config/urls.py
"""
URL configuration for config project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/5.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import include, path
from django.conf.urls.static import static
from health_check.views import HealthCheckView

from django.conf import settings



urlpatterns = [    
    path('', include(('main.urls','main'),namespace='main')),
    path('admin/', admin.site.urls),
    path('users/', include(('users.urls','users'), namespace='users')),   
    path('users/', include('django.contrib.auth.urls')),
    path('shop/', include(('shop.urls','shop'), namespace='shop')),
    path("geo/", include(("geo.urls",'geo'), namespace="geo")),
    path('health/', HealthCheckView.as_view()),
]

if settings.DEBUG:
    import debug_toolbar
    urlpatterns = [
        path('__debug__/', include(debug_toolbar.urls)),
    ] + urlpatterns
    
    # Serve static and media files in development
    urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)


handler400 = "config.views.custom_400"
handler403 = "config.views.custom_403"
handler404 = "config.views.custom_404"
handler500 = "config.views.custom_500"
