requirements.txt
asgiref==3.9.1
black==26.3.1
certifi==2025.11.12
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cssbeautifier==1.15.4
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.13
django-countries==7.6.1
django-csp==4.0
django-debug-toolbar==6.0.0
django-health-check==4.0.0
django-redis==6.0.0
django-widget-tweaks==1.5.0
djlint==1.36.4
dnspython==2.8.0
EditorConfig==0.17.1
filelock==3.20.3
gunicorn==25.0.1
hashids==1.3.1
hiredis==3.3.0
identify==2.6.15
idna==3.11
jsbeautifier==1.15.4
json5==0.12.1
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==25.0
pathspec==1.0.0
pillow==12.2.0
platformdirs==4.4.0
pre_commit==4.5.1
psycopg==3.2.12
psycopg-binary==3.2.12
python-decouple==3.8
python-dotenv==1.2.2
python-slugify==8.0.4
pytokens==0.4.0
PyYAML==6.0.3
redis==7.1.0
regex==2025.11.3
requests==2.33.0
six==1.17.0
sqlparse==0.5.4
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
virtualenv==20.36.1
